feat: Remove UserContainer [BLOX-159]
